+++ src/hooks/common/useScrollToTop.ts
@@ -1,0 +1,7 @@
+import { useEffect } from 'react'
+
+export default function useScrollToTop() {
+    useEffect(() => {
+		window.scrollTo(0, 0);
+	}, []);
+}
